+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Timeout parameter to ReadRawAccelerationsFromFifo and ReadAccelerationsFromFifo.
+- Tests for ReadStatus, ClearFifo, ReadAccelerations, ReadTemperature and SelfTest output values.
 
 ### Fixed
 - SelfTest not restoring the acceleration range.